Made S3_SRC a required env var

+++ chuckit.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
-
+if [ -z "$S3_SRC" ]; then
+    echo "Need to set S3_SRC! e.g. export S3_SRC=s3://bucket-name/foo/dollar/"
+    exit 1
+fi
 echo '000 cleaning up 000'
 echo '$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$'
 rm tmp/*
@@ -19,7 +22,6 @@ for file in tmp/*; do echo "moving '${file}' --> 'tmp/x12005010x220a1834out-${TS
 for file in tmp/*; do mv "$file" "tmp/x12005010x220a1834out-${TSTAMP}-${file##*/}"; done
 echo '~~~~~~~~~><CHUCKING!!!!'
 #export S3_SRC=s3://grnds-development-imports/walmartdemo/dollar/$(date +%s)
-export S3_SRC=s3://grnds-development-imports/walmartdemo/dollar/
 echo '$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$'
 echo "...files landed in ${S3_SRC}"
 aws s3 cp tmp/ $S3_SRC --recursive
